+++ mrv2/docs/es/_static/documentation_options.js
@@ -1,5 +1,5 @@
 const DOCUMENTATION_OPTIONS = {
-    VERSION: 'v1.1.8',
+    VERSION: 'v1.1.9',
     LANGUAGE: 'es',
     COLLAPSE_INDEX: false,
     BUILDER: 'html',